requirements.txt
Flask==3.1.1
flask_cors==6.0.0
frenztoolkit==0.2.5
joblib==1.4.2
keras==3.10.0
matplotlib==3.9.4
mne==1.9.0
mne_connectivity==0.7.0
numpy==2.2.6
pandas==2.2.3
psutil==6.0.0
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sounddevice==0.5.1
soundfile==0.13.1
tensorflow==2.19.0
websockets==15.0.1
